add endpoint to create some products

+++ src/main/java/com/superapp/product/ProductDto.java
@@ -1,9 +1,5 @@
 package com.superapp.product;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -15,5 +11,5 @@ public class ProductDto {
     private Integer id;
     private String name;
     private String description;
-    private Integer unitPrice;
+    private Float unitPrice;
 }
